Installed & configured pinentry for password-store and auth-source.
This depends on a wrapper /usr/bin/pinentry-emacs being used as the
pinentry program. In order to use pinentry outside of emacs, a wrapper
script is used on the system to switch the pinentry program depending
on what the environment variable PINENTRY_USER_DATA is set to.

Reference: https://a3nm.net/git/mybin/file/my-pinentry.html